apply nuno's idea to use unsafe.Pointer

+++ internal/engine/compiler/impl_arm64_test.go
@@ -47,7 +47,7 @@ func TestArm64Compiler_indirectCallWithTargetOnCallingConvReg(t *testing.T) {
 			moduleInstance:     env.moduleInstance,
 		}
 		me.functions = append(me.functions, f)
-		table[0] = uintptr(unsafe.Pointer(&f))
+		table[0] = unsafe.Pointer(&f)
 	}
 
 	compiler := env.requireNewCompiler(t, &wasm.FunctionType{}, newCompiler, &wazeroir.CompilationResult{
